pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.84.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "google-generativeai>=0.8.5",
    "trafilatura>=2.0.0",
    "firebase-admin>=6.9.0",
    "pyrebase4>=4.8.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
